+++ packages/apollo-server-cloud-function/src/googleCloudApollo.ts
@@ -18,7 +18,8 @@ export function graphqlCloudFunction(options: GraphQLOptions): any {
   }
 
   const graphqlHandler: any = (req: Request, res: Response): void => {
-    if (req.method === 'POST' && !req.body) {
+    const hasPostBody = req.body && Object.keys(req.body).length > 0;
+    if (req.method === 'POST' && !hasPostBody) {
       res.status(500).send('POST body missing.');
       return;
     }
@@ -26,7 +27,7 @@ export function graphqlCloudFunction(options: GraphQLOptions): any {
     runHttpQuery([req, res], {
       method: req.method,
       options: options,
-      query: req.method === 'POST' ? req.body : (req.query as any),
+      query: hasPostBody ? req.body : (req.query as any),
       request: {
         url: req.url,
         method: req.method,
@@ -40,14 +41,10 @@ export function graphqlCloudFunction(options: GraphQLOptions): any {
           .send(graphqlResponse);
       },
       (error: HttpQueryError) => {
-        console.log('Error!');
-        console.log(JSON.stringify(error));
         if ('HttpQueryError' !== error.name) {
           res.status(500).send(error);
           return;
         }
-        console.log('other error');
-        console.log(JSON.stringify(error));
         res
           .status(error.statusCode)
           .set(error.headers)