+++ packages/apollo-server-env/src/fetch.d.ts
@@ -1,3 +1,5 @@
+import { Agent } from 'http';
+
 export declare function fetch(
   input?: RequestInfo,
   init?: RequestInit,
@@ -50,6 +52,14 @@ export interface RequestInit {
   referrer?: string;
   referrerPolicy?: ReferrerPolicy;
   integrity?: string;
+
+  // The following properties are node-fetch extensions
+  follow?: number;
+  timeout?: number;
+  compress?: boolean;
+  size?: number;
+  agent?: Agent;
+
   // Cloudflare Workers accept a `cf` property to control Cloudflare features
   // See https://developers.cloudflare.com/workers/reference/cloudflare-features/
   cf?: {